Rename package: hexagonaldocumentation
