+++ src/main/java/tech/jhipster/lite/generator/client/hexagonaldocumentation/application/FrontHexagonalArchitectureDocumentationApplicationService.java
@@ -1,7 +1,7 @@
-package tech.jhipster.lite.generator.client.hexagonal_documentation.application;
+package tech.jhipster.lite.generator.client.hexagonaldocumentation.application;
 
 import org.springframework.stereotype.Service;
-import tech.jhipster.lite.generator.client.hexagonal_documentation.domain.FrontHexagonalDocumentationModuleFactory;
+import tech.jhipster.lite.generator.client.hexagonaldocumentation.domain.FrontHexagonalDocumentationModuleFactory;
 import tech.jhipster.lite.module.domain.JHipsterModule;
 import tech.jhipster.lite.module.domain.properties.JHipsterModuleProperties;
